+++ src/tests/test.index.js
@@ -1,6 +1,6 @@
 const assert = require('chai').assert;
 const format = require('../index.js');
-const BN = require('bn.js');
+const BN = require('bignumber.js');
 const schema = require('../schema.json');
 
 describe('test ethjs-format object', () => {
@@ -16,6 +16,7 @@ describe('test ethjs-format object', () => {
       const emptyString = "";
       const zeroNumber = 0;
       const justPrefix = "0x";
+      const floatNumber = "238428.237842";
 
       assert.equal(format.formatQuantity(emptyString, true), '0x0');
       assert.equal(format.formatQuantity(zeroNumber, true), '0x0');
@@ -27,6 +28,12 @@ describe('test ethjs-format object', () => {
       assert.equal(format.formatQuantity(noPrefixHexNumber, true), '0x21D21A2'.toLowerCase());
       assert.equal(format.formatQuantity(prefixHexNumber, true), '0x21D21A2'.toLowerCase());
       assert.equal(format.formatQuantity(justPrefix, true), '0x0');
+
+      try {
+        format.formatQuantity(floatNumber, true);
+      } catch (error) {
+        assert.equal(typeof error, 'object');
+      }
     });
 
     it('should decode normally', () => {
@@ -41,6 +48,7 @@ describe('test ethjs-format object', () => {
       const emptyString = "";
       const zeroNumber = 0;
       const justPrefix = '0x';
+      const floatNumber = "238428.237842";
 
       const r1 = format.formatQuantity(prefixHexNumber, false).toString(10);
       const r2 = format.formatQuantity(noPrefixHexNumber, false).toString(10);
@@ -64,6 +72,12 @@ describe('test ethjs-format object', () => {
       assert.equal(r9, '0');
       assert.equal(r10, '0');
       assert.equal(format.formatQuantity(undefinedNumber, false), undefinedNumber);
+
+      try {
+        format.formatQuantity(floatNumber, false);
+      } catch (error) {
+        assert.equal(typeof error, 'object');
+      }
     });
   });
 
@@ -142,8 +156,8 @@ describe('test ethjs-format object', () => {
         logs: [{
           "logIndex": "0x1", // 1
           "blockNumber":"0x1b4", // 436
-          "blockHash": "0x8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcfdf829c5a142f1fccd7d",
-          "transactionHash":  "0xdf829c5a142f1fccd7d8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcf",
+          "blockHash": "0x8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcfdf829c5a142f1fccd7d31",
+          "transactionHash":  "0xdf829c5a142f1fccd7d8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcf55",
           "transactionIndex": "0x0", // 0
           "address": "0x16c5785ac562ff41e2dcfdf829c5a142f1fccd7d",
           "data":"0x0000000000000000000000000000000000000000000000000000000000000000",
@@ -166,8 +180,8 @@ describe('test ethjs-format object', () => {
       assert.equal(decodedObject.logs[0].data, '0x0000000000000000000000000000000000000000000000000000000000000000');
       assert.equal(decodedObject.logs[0].data, '0x0000000000000000000000000000000000000000000000000000000000000000');
       assert.equal(decodedObject.logs[0].address, '0x16c5785ac562ff41e2dcfdf829c5a142f1fccd7d');
-      assert.equal(decodedObject.logs[0].blockHash, '0x8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcfdf829c5a142f1fccd7d');
-      assert.equal(decodedObject.logs[0].transactionHash, '0xdf829c5a142f1fccd7d8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcf');
+      assert.equal(decodedObject.logs[0].blockHash, '0x8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcfdf829c5a142f1fccd7d31');
+      assert.equal(decodedObject.logs[0].transactionHash, '0xdf829c5a142f1fccd7d8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcf55');
 
       assert.equal(decodedObject.logs[0].logIndex.toString(10), '1');
       assert.equal(decodedObject.logs[0].blockNumber.toString(10), '436');
@@ -182,7 +196,7 @@ describe('test ethjs-format object', () => {
         "blockNumber": "0x15df", // 5599
         "transactionIndex":  "0x1", // 1
         "from":"0x407d73d8a49eeb85d32cf465507dd71d507100c1",
-        "to":"0x85h43d8a49eeb85d32cf465507dd71d507100c1",
+        "to":"0x85h43d8a49eeb85d32cf465507dd71d507100c12",
         "value":"0x7f110", // 520464
         "gas": "0x7f110", // 520464
         "gasPrice":"0x09184e72a000",
@@ -200,7 +214,7 @@ describe('test ethjs-format object', () => {
       assert.equal(decodedObject.value.toString(10), '520464');
       assert.equal(decodedObject.gas.toString(10), '520464');
       assert.equal(decodedObject.from, '0x407d73d8a49eeb85d32cf465507dd71d507100c1');
-      assert.equal(decodedObject.to, '0x85h43d8a49eeb85d32cf465507dd71d507100c1');
+      assert.equal(decodedObject.to, '0x85h43d8a49eeb85d32cf465507dd71d507100c12');
       assert.equal(decodedObject.input, '0x603880600c6000396000f300603880600c6000396000f3603880600c6000396000f360');
     });
 
@@ -212,7 +226,7 @@ describe('test ethjs-format object', () => {
         "blockNumber": "0x15df", // 5599
         "transactionIndex":  "0x1", // 1
         "from":"0x407d73d8a49eeb85d32cf465507dd71d507100c1",
-        "to":"0x85h43d8a49eeb85d32cf465507dd71d507100c1",
+        "to":"0x85h43d8a49eeb85d32cf465507dd71d507100c12",
         "value":"0x7f110", // 520464
         "gas": "0x7f110", // 520464
         "gasPrice":"0x09184e72a000",
@@ -221,11 +235,11 @@ describe('test ethjs-format object', () => {
 
       const encodedBlockObject = {
         "number": "0x1b4", // 436
-        "hash": "0xe670ec64341771606e55d6b4ca35a1a6b75ee3d5145a99d05921026d1527331",
+        "hash": "0x8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcfdf829c5a142f1fccd7d32",
         "parentHash": "0x9646252be9520f6e71339a8df9c55e4d7619deeb018d2a3f2d21fc165dde5eb5",
         "nonce": "0xe04d296d2460cfb8472af2c5fd05b5a214109c25688d3704aed5484f9a7792f2",
         "sha3Uncles": "0x1dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347",
-        "logsBloom": "0xe670ec64341771606e55d6b4ca35a1a6b75ee3d5145a99d05921026d1527331",
+        "logsBloom": "0x8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcfdf829c5a142f1fccd7d32",
         "transactionsRoot": "0x56e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421",
         "stateRoot": "0xd5855eb08b3387c0af375e9cdb6acfc05eb8f519e419b874b6ff2ffda7ed1dff",
         "miner": "0x4e65fda2159562a496f9f3522f89122a3088497a",
@@ -236,17 +250,17 @@ describe('test ethjs-format object', () => {
         "gasLimit": "0x9f759", // 653145
         "gasUsed": "0x9f759", // 653145
         "timestamp": "0x54e34e8e", // 1424182926
-        "transactions": ["0xe670ec64341771606e55d6b4ca35a1a6b75ee3d5145a99d05921026d1527331", encodedTransactionObject],
+        "transactions": ["0x8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcfdf829c5a142f1fccd7d32", encodedTransactionObject],
         "uncles": ["0x9646252be9520f6e71339a8df9c55e4d7619deeb018d2a3f2d21fc165dde5eb5", "0x9646252be9520f6e71339a8df9c55e4d7619deeb018d2a3f2d21fc165dde5eb5"],
       };
 
       const decodedObject = format.formatObject('Block', encodedBlockObject, false);
 
-      assert.equal(decodedObject.hash, '0xe670ec64341771606e55d6b4ca35a1a6b75ee3d5145a99d05921026d1527331');
+      assert.equal(decodedObject.hash, '0x8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcfdf829c5a142f1fccd7d32');
       assert.equal(decodedObject.parentHash, '0x9646252be9520f6e71339a8df9c55e4d7619deeb018d2a3f2d21fc165dde5eb5');
       assert.equal(decodedObject.nonce, '0xe04d296d2460cfb8472af2c5fd05b5a214109c25688d3704aed5484f9a7792f2');
       assert.equal(decodedObject.sha3Uncles, '0x1dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347');
-      assert.equal(decodedObject.logsBloom, '0xe670ec64341771606e55d6b4ca35a1a6b75ee3d5145a99d05921026d1527331');
+      assert.equal(decodedObject.logsBloom, '0x8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcfdf829c5a142f1fccd7d32');
       assert.equal(decodedObject.transactionsRoot, '0x56e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421');
       assert.equal(decodedObject.stateRoot, '0xd5855eb08b3387c0af375e9cdb6acfc05eb8f519e419b874b6ff2ffda7ed1dff');
       assert.equal(decodedObject.miner, '0x4e65fda2159562a496f9f3522f89122a3088497a');
@@ -262,7 +276,7 @@ describe('test ethjs-format object', () => {
       assert.equal(Array.isArray(decodedObject.transactions), true);
       assert.equal(Array.isArray(decodedObject.uncles), true);
 
-      assert.equal(decodedObject.transactions[0], "0xe670ec64341771606e55d6b4ca35a1a6b75ee3d5145a99d05921026d1527331");
+      assert.equal(decodedObject.transactions[0], "0x8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcfdf829c5a142f1fccd7d32");
       assert.equal(decodedObject.uncles[0], "0x9646252be9520f6e71339a8df9c55e4d7619deeb018d2a3f2d21fc165dde5eb5");
 
       assert.equal(decodedObject.transactions[1].hash, '0xc6ef2fc5426d6ad6fd9e2a26abeab0aa2411b7ab17f30a99d3cb96aed1d1055b');
@@ -274,7 +288,7 @@ describe('test ethjs-format object', () => {
       assert.equal(decodedObject.transactions[1].value.toString(10), '520464');
       assert.equal(decodedObject.transactions[1].gas.toString(10), '520464');
       assert.equal(decodedObject.transactions[1].from, '0x407d73d8a49eeb85d32cf465507dd71d507100c1');
-      assert.equal(decodedObject.transactions[1].to, '0x85h43d8a49eeb85d32cf465507dd71d507100c1');
+      assert.equal(decodedObject.transactions[1].to, '0x85h43d8a49eeb85d32cf465507dd71d507100c12');
       assert.equal(decodedObject.transactions[1].input, '0x603880600c6000396000f300603880600c6000396000f3603880600c6000396000f360');
     });
 
@@ -578,6 +592,32 @@ describe('test ethjs-format object', () => {
       assert.equal(encodeNull10, null);
     });
 
+    it ('should handle 20 and 32 byte data properly', () => {
+      assert.equal(format.format('DATA32', '0x9646252be9520f6e71339a8df9c55e4d7619deeb018d2a3f2d21fc165dde5eb5', false), '0x9646252be9520f6e71339a8df9c55e4d7619deeb018d2a3f2d21fc165dde5eb5');
+      assert.equal(format.format('DATA32', '0x8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcfdf829c5a142f1fccd7d32', true), '0x8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcfdf829c5a142f1fccd7d32');
+      assert.equal(format.format('DATA20', '0x407d73d8a49eeb85d32cf465507dd71d507100c1', true), '0x407d73d8a49eeb85d32cf465507dd71d507100c1');
+      assert.equal(format.format('DATA20', '0x85h43d8a49eeb85d32cf465507dd71d507100c12', false), '0x85h43d8a49eeb85d32cf465507dd71d507100c12');
+      assert.equal(format.format('DATA20', '0x', false), '0x');
+      var invalidBytesError;
+
+      try {
+        format.format('DATA20', '*((*(dsfjj)))', false);
+      } catch (errNonAlpha) {
+        assert.equal(typeof errNonAlpha, 'object');
+      }
+
+      try {
+        format.format('DATA32', '0x0', true);
+        format.format('DATA32', '0x89iusdf', false);
+        format.format('DATA20', '0x038372', true);
+        format.format('DATA20', '0x038372', false);
+      } catch (error) {
+        invalidBytesError = error;
+      }
+
+      assert.equal(typeof invalidBytesError, 'object');
+    });
+
     it ('should decode null normally', () => {
       const decodedNull1 = format.format('SendTransaction', null, false);
       const decodedNull2 = format.format('Filter', null, false);
@@ -704,8 +744,8 @@ describe('test ethjs-format object', () => {
       const decodedFilterChanges = format.formatOutputs('eth_getFilterChanges', [{
         "logIndex": "0x1", // 1
         "blockNumber":"0x1b4", // 436
-        "blockHash": "0x8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcfdf829c5a142f1fccd7d",
-        "transactionHash":  "0xdf829c5a142f1fccd7d8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcf",
+        "blockHash": "0x8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcfdf829c5a142f1fccd7d21",
+        "transactionHash":  "0xdf829c5a142f1fccd7d8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcf55",
         "transactionIndex": "0x0", // 0
         "address": "0x16c5785ac562ff41e2dcfdf829c5a142f1fccd7d",
         "data":"0x0000000000000000000000000000000000000000000000000000000000000000",
@@ -713,8 +753,8 @@ describe('test ethjs-format object', () => {
       }, {
         "logIndex": "0x1", // 1
         "blockNumber":"0x1b4", // 436
-        "blockHash": "0x8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcfdf829c5a142f1fccd7d",
-        "transactionHash":  "0xdf829c5a142f1fccd7d8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcf",
+        "blockHash": "0x8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcfdf829c5a142f1fccd7d21",
+        "transactionHash":  "0xdf829c5a142f1fccd7d8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcf55",
         "transactionIndex": "0x0", // 0
         "address": "0x16c5785ac562ff41e2dcfdf829c5a142f1fccd7d",
         "data":"0x0000000000000000000000000000000000000000000000000000000000000000",
@@ -726,8 +766,8 @@ describe('test ethjs-format object', () => {
       assert.equal(decodedFilterChanges[0].transactionIndex.toString(10), '0');
       assert.equal(decodedFilterChanges[0].address, '0x16c5785ac562ff41e2dcfdf829c5a142f1fccd7d');
       assert.equal(decodedFilterChanges[0].data, '0x0000000000000000000000000000000000000000000000000000000000000000');
-      assert.equal(decodedFilterChanges[0].blockHash, '0x8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcfdf829c5a142f1fccd7d');
-      assert.equal(decodedFilterChanges[0].transactionHash, '0xdf829c5a142f1fccd7d8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcf');
+      assert.equal(decodedFilterChanges[0].blockHash, '0x8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcfdf829c5a142f1fccd7d21');
+      assert.equal(decodedFilterChanges[0].transactionHash, '0xdf829c5a142f1fccd7d8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcf55');
       assert.equal(decodedFilterChanges[0].topics[0], '0x59ebeb90bc63057b6515673c3ecf9438e5058bca0f92585014eced636878c9a5');
 
       assert.equal(decodedFilterChanges[1].logIndex.toString(10), '1');
@@ -735,8 +775,8 @@ describe('test ethjs-format object', () => {
       assert.equal(decodedFilterChanges[1].transactionIndex.toString(10), '0');
       assert.equal(decodedFilterChanges[1].address, '0x16c5785ac562ff41e2dcfdf829c5a142f1fccd7d');
       assert.equal(decodedFilterChanges[1].data, '0x0000000000000000000000000000000000000000000000000000000000000000');
-      assert.equal(decodedFilterChanges[1].blockHash, '0x8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcfdf829c5a142f1fccd7d');
-      assert.equal(decodedFilterChanges[1].transactionHash, '0xdf829c5a142f1fccd7d8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcf');
+      assert.equal(decodedFilterChanges[1].blockHash, '0x8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcfdf829c5a142f1fccd7d21');
+      assert.equal(decodedFilterChanges[1].transactionHash, '0xdf829c5a142f1fccd7d8216c5785ac562ff41e2dcfdf5785ac562ff41e2dcf55');
       assert.equal(decodedFilterChanges[1].topics[0], '0x59ebeb90bc63057b6515673c3ecf9438e5058bca0f92585014eced636878c9a5');
     });
   });