+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.0.2 -- ethjs-format
+
+1. handle quantity floats with error (no floats on chain)
+2. switched all bignumbers from `bn.js` to `bignumber.js`
+3. enfore 20 and 32 byte lengths where required, throw if not alphanumeric
+
 # 0.0.1 -- ethjs-formmat
 
 1. Basic testing